requirements.txt
numpy == 1.23.3
opencv-python==4.6.0.66
pillow == 9.2.0
pycocotools == 2.0.6
scipy == 1.9.3
torchvision == 0.13.1
tqdm == 4.64.1
matplotlib == 3.6.2
h5py == 3.7.0 